+++ test/test_stdio.cpp
@@ -8,26 +8,18 @@
 //
 
 #include <boost/nowide/cstdio.hpp>
+
+#include <boost/nowide/convert.hpp>
 #include <cstdlib>
 #include <cstring>
 #include <iostream>
 
 #include "test.hpp"
 
-#ifdef _MSC_VER
-FILE* wfopen(const wchar_t* filename, const wchar_t* mode)
-{
-    FILE* f;
-    return (_wfopen_s(&f, filename, mode) == 0) ? f : NULL;
-}
-#else
-#define wfopen _wfopen
-#endif
-
 bool file_exists(const std::string& filename)
 {
 #ifdef BOOST_WINDOWS
-    FILE* f = wfopen(boost::nowide::widen(filename).c_str(), L"r");
+    FILE* f = boost::nowide::detail::wfopen(boost::nowide::widen(filename).c_str(), L"r");
 #else
     FILE* f = std::fopen(filename.c_str(), "r");
 #endif
@@ -43,7 +35,7 @@ bool file_exists(const std::string& filename)
 void create_test_file(const std::string& filename)
 {
 #ifdef BOOST_WINDOWS
-    FILE* f = wfopen(boost::nowide::widen(filename).c_str(), L"w");
+    FILE* f = boost::nowide::detail::wfopen(boost::nowide::widen(filename).c_str(), L"w");
 #else
     FILE* f = std::fopen(filename.c_str(), "w");
 #endif
